Fixes 238 for srcset
The getBestSelectionFromSrcset function was just returning the smallest
imge if there was no image larger than the container. I changed the loop
to order by largest first and then look to see if the next iteration is
smaller than the container, if so, this is as small as we should go.